Update database from UI (#18)
* [refactor] No more `/api/exams` route

Setting the database.ts file to only server actions using `'use server';` and passing the `getAllExams` function to `async`, we can now directly use the `getAllExams` function directly into the `FullCalendar` component without using an API route.

* [feature] Update event dates from UI

When moving events (exams) on the calendar, it now updates correctly in the database.

The default date if the exam doesn't have the property `crep_print_date` set is now the date at runtime.

Also removed the `All day` slot.

* [feature] Display more options in the Modal's select if the user is admin

* [fix] added missing import

`updateExamById` was not imported, god knows why... my bad

* [fix] Hours & minutes in modal

In the modal, hours and minutes were wrong, with a delay of two hours. It's now fixed.

* [refactor] `updateExamById` → `updateExamDateById`

Since we will need to update other things than the date of the exam, I renamed the function for more explicity.

* [feature] Update exam status using the select

When selecting a new value in the select for the status of an exam, it really updates the status in the database.

* [feature] Remarks in modal working in database

When changing remarks in the modal, it updates correctly in the database.