style: new design (#19)
* feat: implement components of calendar, navbar and authentication buttons

* Use the database for calendar (#16)

* [feature] `/api/exams` route

API route using the `getAllExams` function to fetch all the exams in the database and return it as JSON.

* [feature] Use exams in calendar

Using the previously created `/api/exams` route, the calendar now uses the exams present in the database.

* [feature] .env.example

* [fix] Wrong table name...

Oops... The right table name to use is `crep`.

* [feature] dev docker-compose

Documentation to come...

* [feature] middleware (#17)

Middleware requires authentication on all routes except `/api/auth` (that is used to log in)

So now, `/` and `/api/exams` can not be accessed if not logged in.

* feat: add Modal component that uses data from the selected exam and a button to open the exam's folder

* style: restructure modal with flex and update overall design

* style: fix close button hover color in dark mode

* fix: disable date and time inputs in Modal component

* feat: add style to fullcalendar buttons

* Update database from UI (#18)

* [refactor] No more `/api/exams` route

Setting the database.ts file to only server actions using `'use server';` and passing the `getAllExams` function to `async`, we can now directly use the `getAllExams` function directly into the `FullCalendar` component without using an API route.

* [feature] Update event dates from UI

When moving events (exams) on the calendar, it now updates correctly in the database.

The default date if the exam doesn't have the property `crep_print_date` set is now the date at runtime.

Also removed the `All day` slot.

* [feature] Display more options in the Modal's select if the user is admin

* [fix] added missing import

`updateExamById` was not imported, god knows why... my bad

* [fix] Hours & minutes in modal

In the modal, hours and minutes were wrong, with a delay of two hours. It's now fixed.

* [refactor] `updateExamById` → `updateExamDateById`

Since we will need to update other things than the date of the exam, I renamed the function for more explicity.

* [feature] Update exam status using the select

When selecting a new value in the select for the status of an exam, it really updates the status in the database.

* [feature] Remarks in modal working in database

When changing remarks in the modal, it updates correctly in the database.

---------

Co-authored-by: Azecko <30987143+Azecko@users.noreply.github.com>
Co-authored-by: Azecko <samikarin24@gmail.com>