[feature] adds production-ready docker files and fixes TS typing
* [fix] All types & TS problems

In prevision of the production Dockerfile and docker-compose, I fixed all the TS related problems.
This is necessary to create a production build with Next.js

* [feature] docker-compose-prod & Dockerfile

docker-compose-prod and Dockerfile for the production version.
This docker-compose is NOT meant to be started on a personal computer, but on the production server using the Ansible present on the [CREP.ops](https://github.com/EPFL-CePro/CREP.ops) repository.