[feature] Authentication required for `/register`
Authentication is now needed to access the `/register` route, but it is not necessary to be apart of any special group.

It is still needed to be apart of the `CREP-access` or the `CREP-admin` group to access the `/` route, but the error if the user is not a member of any of those groups is now on the `/403` page and not a direct error returned by Entra.