[feature] Insert exam algorithm
First version of the "algorithm" to calculate the printing date of an exam.
It calculates the necessary duration for an exam to be printed, and choose an available time between the moment were the form is submitted and the desired delivery date (actually, one day before the delivery date)

This will probably change a lot in the future, and the code will be refactored during time.